+++ ibm/provider/provider.go
@@ -390,6 +390,9 @@ func Provider() *schema.Provider {
 			"ibm_is_subnet_reserved_ip":              vpc.DataSourceIBMISReservedIP(),
 			"ibm_is_subnet_reserved_ips":             vpc.DataSourceIBMISReservedIPs(),
 			"ibm_is_security_group":                  vpc.DataSourceIBMISSecurityGroup(),
+			"ibm_is_security_groups":                 vpc.DataSourceIBMIsSecurityGroups(),
+			"ibm_is_security_group_rule":             vpc.DataSourceIBMIsSecurityGroupRule(),
+			"ibm_is_security_group_rules":            vpc.DataSourceIBMIsSecurityGroupRules(),
 			"ibm_is_security_group_target":           vpc.DataSourceIBMISSecurityGroupTarget(),
 			"ibm_is_security_group_targets":          vpc.DataSourceIBMISSecurityGroupTargets(),
 			"ibm_is_snapshot":                        vpc.DataSourceSnapshot(),
