+++ ibm/service/vpc/data_source_ibm_is_security_group_rule_test.go
@@ -1,12 +1,13 @@
 // Copyright IBM Corp. 2021 All Rights Reserved.
 // Licensed under the Mozilla Public License v2.0
 
-package ibm
+package vpc_test
 
 import (
 	"fmt"
 	"testing"
 
+	acc "github.com/IBM-Cloud/terraform-provider-ibm/ibm/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
@@ -15,8 +16,8 @@ func TestAccIBMIsSecurityGroupRuleDataSourceBasic(t *testing.T) {
 	vpcname := fmt.Sprintf("tfsecgrprl-vpc-%d", acctest.RandIntRange(10, 100))
 	sgname := fmt.Sprintf("tfsecgrprl-name-%d", acctest.RandIntRange(10, 100))
 	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acc.TestAccPreCheck(t) },
+		Providers: acc.TestAccProviders,
 		Steps: []resource.TestStep{
 			resource.TestStep{
 				Config: testAccCheckIBMIsSecurityGroupRuleDataSourceConfigBasic(vpcname, sgname),
