+++ ibm/service/vpc/data_source_ibm_is_security_groups.go
@@ -1,7 +1,7 @@
 // Copyright IBM Corp. 2021 All Rights Reserved.
 // Licensed under the Mozilla Public License v2.0
 
-package ibm
+package vpc
 
 import (
 	"context"
@@ -10,12 +10,14 @@ import (
 	"reflect"
 	"time"
 
+	"github.com/IBM-Cloud/terraform-provider-ibm/ibm/conns"
+	"github.com/IBM-Cloud/terraform-provider-ibm/ibm/flex"
 	"github.com/IBM/vpc-go-sdk/vpcv1"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func dataSourceIBMIsSecurityGroups() *schema.Resource {
+func DataSourceIBMIsSecurityGroups() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceIBMIsSecurityGroupsRead,
 
@@ -350,7 +352,7 @@ func dataSourceIBMIsSecurityGroups() *schema.Resource {
 }
 
 func dataSourceIBMIsSecurityGroupsRead(context context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	vpcClient, err := meta.(ClientSession).VpcV1API()
+	vpcClient, err := meta.(conns.ClientSession).VpcV1API()
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -386,7 +388,7 @@ func dataSourceIBMIsSecurityGroupsRead(context context.Context, d *schema.Resour
 			return diag.FromErr(fmt.Errorf("ListSecurityGroupsWithContext failed %s\n%s", err, response))
 		}
 
-		start = GetNext(securityGroupCollection.Next)
+		start = flex.GetNext(securityGroupCollection.Next)
 		allrecs = append(allrecs, securityGroupCollection.SecurityGroups...)
 
 		if start == "" {
